カラオケ検索ダイアログをフルスクリーン化
