+++ app/javascript/mixin/DialogComponentMixin.js
@@ -1,5 +1,6 @@
 //
 // ダイアログコンポーネント共通の仕組み
+// TODO: PCでは500px、スマホではフルスクリーンになるようにここで一元化したい
 //
 export default {
   data: function() {
@@ -12,11 +13,13 @@ export default {
       handler: function(v) {
         if (v) {
           document.getElementsByTagName("body")[0].className="noscroll"
-        } else {
-          document.body.removeAttribute("class","noscroll")
         }
       },
       immediate: true
     }
+  },
+  beforeDestroy: function() {
+    this.isShow = false
+    document.body.removeAttribute("class","noscroll")
   }
 }
